source/silver/pyproject.toml
[project]
name = "opengeh_silver"
version = "0.1.0"
description = "Streaming to the silver layer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opengeh-spark-sql-migrations @ git+https://git@github.com/Energinet-DataHub/opengeh-python-packages@3.0.1#subdirectory=source/spark_sql_migrations",
    "opengeh-telemetry @ git+https://git@github.com/Energinet-DataHub/opengeh-python-packages@3.0.3#subdirectory=source/telemetry",
    "pydantic-settings>=2.7.1",
]

[project.scripts]
stream_calculated_measurements = "opengeh_silver.entry_points:stream_calculated_measurements"
migrate = "opengeh_silver.entry_points:migrate"

[dependency-groups]
dev = [
    "commitizen>=4.1.0",
    "mypy>=1.14.1",
    "opengeh-testcommon",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.9.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
# -s ensures that print("<text>") is written to output. 
addopts = "--cov-report=term --cov-report=xml:coverage.xml --cov=opengeh_silver -ra -vv -s"

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.ruff]
line-length = 120
show-fixes = true

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["D", "D401", "E", "F", "I", "T201", "UP"]
ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D106",
    "D107",
    "D417",
    "E501",
    "UP006",
    "UP007",
    "UP035",
]
exclude = ["notebooks/*"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.uv.sources.opengeh-testcommon]
git = "https://github.com/Energinet-DataHub/opengeh-python-packages"
subdirectory = "source/testcommon"
rev = "3.0.1"
